chore(frontend): package restructure
+ issue#25 refactured to a clean project directory.
+ added dummy buttons in preparation for lobby features.
