--- apps/android/app/src/main/java/at/aau/serg/android/ui/screens/HomeScreen.kt
@@ -1,29 +1,0 @@
-package at.aau.serg.android.ui.screens
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun HomeScreen(
-    modifier: Modifier = Modifier,
-    onShowLeaderboard: () -> Unit
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text("Welcome to the App!")
-
-        Spacer(Modifier.height(16.dp))
-
-        Button(onClick = onShowLeaderboard) {
-            Text("Open Leaderboard")
-        }
-    }
-}
